api/requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.6
django-cors-headers==3.5.0
django-filter==2.3.0
django-mongoengine==0.4.2
django-rest-framework-mongoengine==3.4.1
django-templated-mail==1.1.1
djangorestframework==3.11.1
djoser==2.0.5
drf-yasg==1.17.1
emoji==0.6.0
fasttext==0.9.2
idna==2.10
importlib-metadata==2.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
joblib==0.16.0
Markdown==3.2.2
MarkupSafe==1.1.1
mongoengine==0.20.0
mysqlclient==2.0.1
nltk==3.5
numpy==1.19.2
packaging==20.4
pandas==1.1.2
pybind11==2.5.0
pymongo==3.11.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
regex==2020.7.14
requests==2.24.0
rest-condition==1.0.3
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scipy==1.5.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
summa==1.2.0
tqdm==4.49.0
uritemplate==3.0.1
urllib3==1.25.10
vaderSentiment==3.3.2
zipp==3.2.0
